+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Removed
+
+- The `Hive` connector has been deleted
+- The `Rok` connector has been deleted
+
 ### [4.4.1] 2023-03-30
 
 ### Changed