+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Removed
 
 - The `Hive` connector has been deleted
+- The `Indexima` connector has been deleted
 - The `Rok` connector has been deleted
 
 ### [4.4.1] 2023-03-30