+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - The `Hive` connector has been deleted
 - The `Indexima` connector has been deleted
 - The `Rok` connector has been deleted
+- The `Lightspeed` connector has been deleted
 
 ### [4.4.1] 2023-03-30